add emanation region + refactor region behavior creation
