+++ module/item/sheets/WitcherRitualSheet.js
@@ -28,7 +28,8 @@ export default class WitcherRitualSheet extends WitcherItemSheet {
                 rect: 'WITCHER.Spell.Square',
                 circle: 'WITCHER.Spell.Circle',
                 cone: 'WITCHER.Spell.Cone',
-                ray: 'WITCHER.Spell.Ray'
+                ray: 'WITCHER.Spell.Ray',
+                emanation: 'WITCHER.Spell.emanation'
             }
         };
     }
