+++ module/setup/queries.js
@@ -6,16 +6,9 @@ import { applyStatusEffectToActor } from '../scripts/statusEffects/applyStatusEf
 
 const system = 'TheWitcherTRPG';
 
-
 async function applyTemporaryItemImprovementsToActor(queryData, { timeout }) {
     let actor = fromUuidSync(queryData.actorUuid);
     actor.applyTemporaryItemImprovements(queryData.effects);
-    return true;
-}
-
-async function createRegionFromTemplates(queryData, { timeout }) {
-    let item = fromUuidSync(queryData.uuid);
-    item.system.regionProperties.createRegionFromTemplateUuids(...queryData.data);
     return true;
 }
 
@@ -34,7 +27,7 @@ async function query(queryData, { timeout }) {
         //Item
         'restoreReliability',
         //Region
-        'createRegionFromTemplateUuids'
+        'addBehaviorsToRegionUuids'
     ];
 
     if (queryData.function in callableFunctions) {
@@ -44,7 +37,8 @@ async function query(queryData, { timeout }) {
 
     if (callableEntityFunctions.includes(queryData.function)) {
         let entity = fromUuidSync(queryData.uuid);
-        entity[queryData.function](...queryData.data);
+        entity[queryData.function]?.(...queryData.data);
+        entity.system[queryData.function]?.(...queryData.data);
         return true;
     }
 
@@ -53,6 +47,5 @@ async function query(queryData, { timeout }) {
 
 export function registerQueries() {
     CONFIG.queries[`${system}.applyTemporaryItemImprovements`] = applyTemporaryItemImprovementsToActor;
-    CONFIG.queries[`${system}.createRegionFromTemplates`] = createRegionFromTemplates;
     CONFIG.queries[`${system}.query`] = query;
 }
